BOAC-1529, move our custom Vue.filters to ./filters file; a cleaner main.ts
