+++ src/main.ts
@@ -1,15 +1,15 @@
+import './filters';
 import 'bootstrap-vue/dist/bootstrap-vue.css';
 import 'bootstrap/dist/css/bootstrap.css';
 import _ from 'lodash';
-import { format as formatDate, parse as parseDate } from 'date-fns';
 import App from './App.vue';
 import axios from 'axios';
 import BootstrapVue from 'bootstrap-vue';
 import router from './router';
 import store from './store';
 import Vue from 'vue';
-import VueHighcharts from 'vue-highcharts';
 import VueAnalytics from 'vue-analytics';
+import VueHighcharts from 'vue-highcharts';
 
 // Allow cookies in Access-Control requests
 axios.defaults.withCredentials = true;
@@ -34,32 +34,6 @@ Vue.use(VueAnalytics, {
   checkDuplicatedScript: true
 });
 
-Vue.filter('date', (dateString: string, format: string = 'MMM dd, YYYY') => {
-  let date = parseDate(dateString);
-  return formatDate(date, format);
-});
-Vue.filter('lowercase', (str: string) => {
-  return str.toLowerCase();
-});
-Vue.filter(
-  'pluralize',
-  (noun: string, count: number, substitutions = {}, pluralSuffix = 's') => {
-    return (
-      `${substitutions[count] || count} ` +
-      (count !== 1 ? `${noun}${pluralSuffix}` : noun)
-    );
-  }
-);
-Vue.filter('round', function(value, decimals) {
-  return (
-    Math.round(value * Math.pow(10, decimals)) / Math.pow(10, decimals)
-  ).toFixed(decimals);
-});
-Vue.filter(
-  'variablePrecisionNumber',
-  (value, minPrecision, maxPrecision) =>
-    `TODO: ${value}, ${minPrecision}, ${maxPrecision}`
-);
 // Emit, and listen for, events via hub
 Vue.prototype.$eventHub = new Vue();
